refactor(insight): 발굴 surface 로직을 공유 recommendDiscoveries로 추출
weekly-verification의 private surfaceDiscoveries를 discovery-recommend.ts의
export recommendDiscoveries로 이동(순수 추출, 동작 동일). 월요일 검증 후 발굴과
후속 데일리 재추천이 같은 surface 함수를 공유하기 위한 함수 레벨 DRY.
